fix: background fd bug
There was a bug that when user quits the extensions, the running
indexing using fd keeps going as it is a child process and gets
reparented to root.

This fix prevents that by piping the output of fd through the extension
so that once the extension stops, the fd exits with failure as the pipe
is broken.